finished urban planning2 exercise
